Cleaner XMLGenDocBuilder code
Designed as a more general approach to concurrent generation of diverse element trees, which didn’t work; the Saxon push generator is inherently serial, which makes sense, I guess. But the code we wrote is a lot cleaner than the code we replaced, so we keep it and it works in the serial case.

We need to have another approach to generating the multiple “phases” of the auction site.