xmlgen - open auctions - connected bid values
The values are indicated in pence; nosqlbench is easier to use (more complete and consistent) using long values, so we use these in our derivation of the bid values.

We use the MVEL internal expressions to relate the bid values, initial value, and current value.
We make the reserve be somewhat bigger than the initial value.

TODO on open auctions - date/time of bids
TODO on open auctions - format longs